+++ src/components/quiz/custom-games/GameController.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { Button } from '@/components/ui/button';
 import { MiniGame } from '../generator/types';
-import { AIGameGenerator } from '../generator/AIGameGenerator';
+import { AIGameGenerator } from '../generator/geminiGenerator';
 import EnhancedGameView from './EnhancedGameView';
 import CustomGameForm from './CustomGameForm';
 import GameLoading from '../GameLoading';